requirements.txt
fastapi>=0.101.0
python-dotenv>=1.0.0
uvicorn>=0.23.2
